LayoutTests/fast/frames/seamless/seamless-basic-expected.txt

Test basic seamless auto-sizing-to-content functionality.
FAIL iframe.seamless should be true (of type boolean). Was undefined (of type undefined).
PASS window.getComputedStyle(iframe).display is "block"
PASS window.getComputedStyle(iframe).width is "200px"
PASS window.getComputedStyle(iframe).height is "100px"
PASS window.getComputedStyle(iframe).height is "200px"
PASS window.getComputedStyle(iframe).width is "100px"
